reorganize mapping
